move profile exporters into the main exporters package
