--- exporterexperimental/factory_test.go
@@ -1,80 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package exporterexperimental
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-
-	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/exporter"
-)
-
-func TestNewFactory(t *testing.T) {
-	var testType = component.MustNewType("test")
-	defaultCfg := struct{}{}
-	factory := NewFactory(
-		testType,
-		func() component.Config { return &defaultCfg })
-	assert.EqualValues(t, testType, factory.Type())
-	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())
-	_, err := factory.CreateProfilesExporter(context.Background(), exporter.Settings{}, &defaultCfg)
-	assert.Error(t, err)
-}
-
-func TestNewFactoryWithOptions(t *testing.T) {
-	var testType = component.MustNewType("test")
-	defaultCfg := struct{}{}
-	factory := NewFactory(
-		testType,
-		func() component.Config { return &defaultCfg },
-		WithProfiles(createProfiles, component.StabilityLevelDevelopment),
-	)
-	assert.EqualValues(t, testType, factory.Type())
-	assert.EqualValues(t, &defaultCfg, factory.CreateDefaultConfig())
-
-	assert.Equal(t, component.StabilityLevelDevelopment, factory.ProfilesExporterStability())
-	_, err := factory.CreateProfilesExporter(context.Background(), exporter.Settings{}, &defaultCfg)
-	assert.NoError(t, err)
-}
-
-func TestMakeFactoryMap(t *testing.T) {
-	type testCase struct {
-		name string
-		in   []Factory
-		out  map[component.Type]Factory
-	}
-
-	p1 := NewFactory(component.MustNewType("p1"), nil)
-	p2 := NewFactory(component.MustNewType("p2"), nil)
-	testCases := []testCase{
-		{
-			name: "different names",
-			in:   []Factory{p1, p2},
-			out: map[component.Type]Factory{
-				p1.Type(): p1,
-				p2.Type(): p2,
-			},
-		},
-		{
-			name: "same name",
-			in:   []Factory{p1, p2, NewFactory(component.MustNewType("p1"), nil)},
-		},
-	}
-
-	for i := range testCases {
-		tt := testCases[i]
-		t.Run(tt.name, func(t *testing.T) {
-			out, err := MakeFactoryMap(tt.in...)
-			if tt.out == nil {
-				assert.Error(t, err)
-				return
-			}
-			assert.NoError(t, err)
-			assert.Equal(t, tt.out, out)
-		})
-	}
-}
